requirements.txt
Flask==1.0.2
Flask-Cors==3.0.6
SQLAlchemy==1.3.0
Flask-SQLAlchemy==2.3.2
gunicorn==19.9.0
mysql-connector==2.1.6